+++ go/pkg/basecamp/client_correspondences.go
@@ -228,7 +228,7 @@ func clientCorrespondenceFromGenerated(gc generated.ClientCorrespondence) Client
 
 	if gc.Creator.Id != 0 || gc.Creator.Name != "" {
 		c.Creator = &Person{
-			ID:           gc.Creator.Id,
+			ID:           int64(gc.Creator.Id),
 			Name:         gc.Creator.Name,
 			EmailAddress: gc.Creator.EmailAddress,
 			AvatarURL:    gc.Creator.AvatarUrl,
