+++ go/pkg/basecamp/messages.go
@@ -476,7 +476,7 @@ func messageFromGenerated(gm generated.Message) Message {
 
 	if gm.Creator.Id != 0 || gm.Creator.Name != "" {
 		m.Creator = &Person{
-			ID:           gm.Creator.Id,
+			ID:           int64(gm.Creator.Id),
 			Name:         gm.Creator.Name,
 			EmailAddress: gm.Creator.EmailAddress,
 			AvatarURL:    gm.Creator.AvatarUrl,
