+++ go/pkg/basecamp/recordings.go
@@ -397,7 +397,7 @@ func recordingFromGenerated(gr generated.Recording) Recording {
 
 	if gr.Creator.Id != 0 || gr.Creator.Name != "" {
 		r.Creator = &Person{
-			ID:           gr.Creator.Id,
+			ID:           int64(gr.Creator.Id),
 			Name:         gr.Creator.Name,
 			EmailAddress: gr.Creator.EmailAddress,
 			AvatarURL:    gr.Creator.AvatarUrl,
