+++ go/pkg/basecamp/todos.go
@@ -570,7 +570,7 @@ func todoFromGenerated(gt generated.Todo) Todo {
 
 	if gt.Creator.Id != 0 || gt.Creator.Name != "" {
 		t.Creator = &Person{
-			ID:           gt.Creator.Id,
+			ID:           int64(gt.Creator.Id),
 			Name:         gt.Creator.Name,
 			EmailAddress: gt.Creator.EmailAddress,
 			AvatarURL:    gt.Creator.AvatarUrl,
